refact: create stub object values
